+++ modules/linux/common/module.c
@@ -63,6 +63,8 @@ VOID __attribute__((constructor)) SymCryptModuleMain(void)
         SymCryptSshKdfSha256SelfTest();
         SymCryptSshKdfSha512SelfTest();
 
+        SymCryptSskdfSelfTest();
+
         SymCryptHmacSha3_256Selftest();
 
         g_SymCryptFipsSelftestsPerformed |= SYMCRYPT_SELFTEST_ALGORITHM_STARTUP;