+++ modules/windows/user/module.c
@@ -59,6 +59,8 @@ PerformStartupAlgorithmSelftests()
     SymCryptSshKdfSha256SelfTest();
     SymCryptSshKdfSha512SelfTest();
 
+    SymCryptSskdfSelfTest();
+
     SymCryptHmacSha3_256Selftest();
 
     g_SymCryptFipsSelftestsPerformed |= SYMCRYPT_SELFTEST_ALGORITHM_STARTUP;