+++ unittest/lib/selftestFuncList.cpp
@@ -78,6 +78,7 @@ const SELFTEST_INFO g_selfTests[] =
     {&SymCryptSrtpKdfSelfTest, "SrtpKdf" },
     {&SymCryptSshKdfSha256SelfTest, "SshKdfSha256" },
     {&SymCryptSshKdfSha512SelfTest, "SshKdfSha512" },
+    {&SymCryptSskdfSelfTest, "Sskdf" },
 
     {NULL, NULL},
 };