+++ CHANGELOG.md
@@ -4,6 +4,7 @@ New changes will be listed here as they are developed. The version number is det
 prior to the creation of a new release, based on the changes contained in that release.
 
 - Internal self-test changes to support FIPS 140-3 certification
+- Add SSKDF implementation
 
 # Version 103.4.3